+++ CHANGELOG.md
@@ -14,6 +14,9 @@ For details about compatibility between different releases, see the **Commitment
 ### Changed
 
 - Low-level log messages from the `go-redis` library are printed only when the log level is set to `DEBUG`.
+- GS will discard repeated gateway uplink messages (often received due to buggy gateway forwarder implementations). A gateway uplink is considered to be repeated when it has the same payload, frequency and antenna index as the last one.
+  - The new `gs_uplink_repeated_total` metric counts how many repeated uplinks have been discarded.
+  - A `gs.up.repeat` event is emitted (once per minute maximum) for gateways that are stuck in a loop and forward the same uplink message.
 
 ### Deprecated